// Sche_Treatment_Project/src/Models/Model.tsx
export interface DoctorEntity {
  id: string;
  accountPhone: string;
  accountEmail: string;
  accountName: string;
  accountGender: number;
  createAt: string;
  updateAt: string;
  supportRole: Support;
  supportStatus: Support;
  doctorDegree: string;
  doctorRank: string;
  doctorSpecialty: string;
  doctorIntroduce: string;
  doctorExp: string;
  doctorImage: string;
}
// ...
export interface CalendarDTO {
  id: string;
  calendarDate: string;
  clinicId: string;
  accountId: string;
  idGroupTime: string;
}
export interface CalendarModel {
  id: string;
  calendarDate: string;
  createAt: string;
  updateAt: string;
  doctor: DoctorEntity;
  groupTime: GroupTime;
}

export interface GroupTime {
  id: string;
  groupTimeDescription: string;
  supports: Support[];
}
export interface Support {
  id: string;
  supportValue: string;
  supportInfo: string;
  createAt: string;
  updateAt: string;
  groupTime: GroupTime;
}